requirements.txt
django
psycopg2
python-environ
django-environ
google-play-scraper
requests
urllib3
app-store-scraper
pandas
openpyxl
transformers
torch 
torchvision
sentencepiece
protobuf
googletrans==3.1.0-alpha
deep-translator
translate
translators
detectlanguage
openai
google-generativeai
joblib
scikit-learn
imblearn
hyperopt
channels
channels_redis
